requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.45.0
pydantic==2.9.1
pydantic-core==2.23.3
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.45.0
pydantic==2.9.1
pydantic-core==2.23.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
